+++ src/pointcloudset/diff/nearest.py
@@ -1,7 +1,8 @@
 """
 Functions to calculate differences between the pointcloud nearest points in another one.
 """
-import numpy as np
+
+from scipy.spatial import KDTree
 
 
 def calculate_distance_to_nearest(pointcloud, target):
@@ -25,8 +26,6 @@ def calculate_distance_to_nearest(pointcloud, target):
     """
     if "distance to nearest point" in pointcloud.data.columns:
         raise ValueError("distance to nearest point already exists.")
-    pc1 = pointcloud.to_instance("open3d")
-    pc2 = target.to_instance("open3d")
-    distance = pc1.compute_point_cloud_distance(pc2)
-    pointcloud._add_column("distance to nearest point", np.array(distance))
+    distances, _ = KDTree(target.points.xyz).query(pointcloud.points.xyz)
+    pointcloud._add_column("distance to nearest point", distances)
     return pointcloud