+++ tests/unit/test_instance_io_pointcloud.py
@@ -1,7 +1,5 @@
 from pathlib import Path
 
-import numpy as np
-import open3d as o3d
 import pandas as pd
 import pytest
 import pytest_check as check
@@ -23,32 +21,11 @@ def test_to_pyntcloud(testpointcloud_mini: PointCloud, testlas1: Path):
     pyntcloud_data = PyntCloud.from_file(testlas1.as_posix())
     res = testpointcloud_mini.to_instance("pyntcloud")
     check.is_instance(res, PyntCloud)
-    check.equal(
-        list(res.points.columns.values), list(testpointcloud_mini.data.columns.values)
-    )
+    check.equal(list(res.points.columns.values), list(testpointcloud_mini.data.columns.values))
 
 
-def test_to_open3d(testpointcloud_mini: PointCloud):
-    pointcloud = testpointcloud_mini.to_instance("open3d")
-    check.equal(type(pointcloud), o3d.geometry.PointCloud)
-    check.equal(pointcloud.dimension(), 3)
-    check.equal(pointcloud.has_points(), True)
-    check.equal(len(np.asarray(pointcloud.points)), len(testpointcloud_mini))
-
-
-def test_from_open3d(testpointcloud_mini_real: PointCloud):
-    open3d_data = testpointcloud_mini_real.to_instance("open3d")
-    pointcloud = pointcloudset.PointCloud.from_instance("open3d", open3d_data)
-    test = pointcloud.data - testpointcloud_mini_real.data[["x", "y", "z"]]
-    check.equal(set(list(test.max())).intersection([0.0, 0.0, 0.0]), {0.0})
-
-
-def test_from_dataframe(
-    testpointcloud_mini_df: pd.DataFrame, testpointcloud_mini: PointCloud
-):
-    pointcloud = pointcloudset.PointCloud.from_instance(
-        "DataFrame", testpointcloud_mini_df
-    )
+def test_from_dataframe(testpointcloud_mini_df: pd.DataFrame, testpointcloud_mini: PointCloud):
+    pointcloud = pointcloudset.PointCloud.from_instance("DataFrame", testpointcloud_mini_df)
     check.is_instance(pointcloud, PointCloud)
     test = pointcloud.data - testpointcloud_mini.data[["x", "y", "z"]]
     check.equal(set(list(test.max())).intersection([0.0, 0.0, 0.0]), {0.0})